move this agendaItem into domain layer so it does not violate dependencies
