+++ app/src/main/java/com/example/tasky/agenda_details/presentation/AgendaDetailsViewModel.kt
@@ -9,9 +9,9 @@ import com.example.tasky.agenda_details.domain.model.Photo
 import com.example.tasky.agenda_details.domain.repository.AgendaDetailsRemoteRepository
 import com.example.tasky.agenda_details.presentation.utils.DateTimeHelper
 import com.example.tasky.common.domain.Result
+import com.example.tasky.common.domain.model.AgendaItemType
 import com.example.tasky.common.presentation.LineItemType
 import com.example.tasky.common.presentation.ReminderTime
-import com.example.tasky.common.presentation.model.AgendaItemType
 import com.example.tasky.common.presentation.util.toFormatted_MMM_dd_yyyy
 import com.example.tasky.feature_agenda.data.model.Attendee
 import com.vanpra.composematerialdialogs.MaterialDialogState
