+++ app/src/main/java/com/example/tasky/common/domain/model/AgendaItemType.kt
@@ -1,4 +1,4 @@
-package com.example.tasky.common.presentation.model
+package com.example.tasky.common.domain.model
 
 enum class AgendaItemType {
     Event, Task, Reminder
